+++ addOns/soap/src/main/java/org/zaproxy/zap/extension/soap/WSDLFilePassiveScanRule.java
@@ -57,11 +57,9 @@ public class WSDLFilePassiveScanRule extends PluginPassiveScanner {
             HttpResponseHeader header = msg.getResponseHeader();
             String baseURL = msg.getRequestHeader().getURI().toString().trim();
             String contentType = header.getHeader(HttpHeader.CONTENT_TYPE).trim();
-            if (baseURL.endsWith(".wsdl")
+            return baseURL.endsWith(".wsdl")
                     || contentType.equals("text/xml")
-                    || contentType.equals("application/wsdl+xml")) {
-                return true;
-            }
+                    || contentType.equals("application/wsdl+xml");
         }
         return false;
     }