+++ addOns/soap/src/test/java/org/zaproxy/zap/extension/soap/SOAPMsgConfigTestCase.java
@@ -31,11 +31,6 @@ import org.junit.jupiter.api.Test;
 
 public class SOAPMsgConfigTestCase {
 
-    private static final String KEY_PREFIX = "xpath:/";
-    private static final String KEY_NAME = "testkey";
-    private static final String KEY_VALUE = "testvalue";
-    private static final String KEY_VALUE2 = "testvalue2";
-
     private SOAPMsgConfig soapConfig;
 
     @BeforeEach
@@ -44,7 +39,7 @@ public class SOAPMsgConfigTestCase {
         soapConfig = new SOAPMsgConfig();
         soapConfig.setWsdl(new Definitions());
         soapConfig.setSoapVersion(1);
-        soapConfig.setParams(new HashMap<String, String>());
+        soapConfig.setParams(new HashMap<>());
         soapConfig.setPort(new Port());
         soapConfig.setBindOp(new BindingOperation());
     }
@@ -67,7 +62,7 @@ public class SOAPMsgConfigTestCase {
 
         soapConfig.setParams(null);
         assertFalse(soapConfig.isComplete()); // Null params.
-        soapConfig.setParams(new HashMap<String, String>());
+        soapConfig.setParams(new HashMap<>());
 
         soapConfig.setPort(null);
         assertFalse(soapConfig.isComplete()); // Null port.
@@ -76,35 +71,5 @@ public class SOAPMsgConfigTestCase {
         soapConfig.setBindOp(null);
         assertFalse(soapConfig.isComplete()); // Null binding operation.
         soapConfig.setBindOp(new BindingOperation());
-    }
-
-    @Test
-    public void changeParamTest() {
-        /* Configuration. */
-        HashMap<String, String> map = new HashMap<String, String>();
-        map.put(KEY_PREFIX + KEY_NAME, KEY_VALUE);
-        soapConfig.setParams(map);
-
-        /* Checks that configuration has been processed correctly. */
-        map = soapConfig.getParams();
-        String value = map.get(KEY_PREFIX + KEY_NAME);
-        assertTrue(value.equals(KEY_VALUE));
-
-        /* Positive case. */
-        soapConfig.changeParam(KEY_NAME, KEY_VALUE2);
-        map = soapConfig.getParams();
-        value = map.get(KEY_PREFIX + KEY_NAME);
-        assertTrue(value.equals(KEY_VALUE2)); // Parameter value has been changed.
-
-        /* Negative cases. */
-        soapConfig.changeParam(KEY_NAME, null); // Null value.
-        map = soapConfig.getParams();
-        value = map.get(KEY_PREFIX + KEY_NAME);
-        assertTrue(value.equals(KEY_VALUE2)); // Parameter value has NOT been changed.
-
-        soapConfig.changeParam(null, KEY_VALUE); // Null key.
-        map = soapConfig.getParams();
-        value = map.get(KEY_PREFIX + KEY_NAME);
-        assertTrue(value.equals(KEY_VALUE2)); // Parameter value has NOT been changed.
     }
 }