+++ addOns/soap/CHANGELOG.md
@@ -10,12 +10,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Change default accelerator for "Import a WSDL file from local file system".
 - Update minimum ZAP version to 2.9.0.
 - Add import menus to (new) top level Import menu instead of Tools menu.
+- Add support for SOAP version 1.2 to the Action Spoofing Scan Rule.
+- Distinguish alerts by adding the SOAP version to the "Other Info" section.
 - Maintenance changes.
 
 ### Fixed
 - Various fixes (related to Issue 4832 and other testing).
 - Fix exception with Java 9+ (Issue 4037).
 - SOAP operations are no longer overwritten in sites tree (Issue 1867).
+- Persist the add-on configuration required by the scan rules in the ZAP database (Issue 4866).
 
 ## 3 - 2017-03-31